refactor: stop using autogenerated gapic chains
Fully compose own callable chains. This will remove the need to fiddle with StubSettings
to disable default retry behavior and will also  eliminate unnecessary gapic trace spans. This also fixes a  bug where the trace annotations for ReadRows was missing the connection id. And this will pave the way for decoupling Opencensus via the ApiTracer api