fix: client.close() should wait until the channels are terminated before shutting down the executor (#916)
* fix: client.close() should wait until the channels are terminated before shutting down the executor

Previously the client.close() would simply shutdown the channel pool and the executor immediately after. Unfortunately this leads to RPCs that would hang forever because an outstanding RPC didnt have an executor to notify of its completion. This PR ensures that the channels are drained before shutting down the executor

* copyright