chore: fix parallel execution and verbose grpc logs
Mucking around with java util logging fails when tests are executed in parallel. So make the 2 features exclusive: by default tests are executed in parallel, if verbose logs are needed they can be activated by a profile that will disable parallel execution.

Also switch parallel tests to run in threads instead of processes and enable parallel execution for unit tests