chore(test): deflake and refactor test fake service helper
There is a race condition when running unit tests in parallel that where they can pick the same port for a fake service. This PR adds a retry loop to let them resolve the race.
It also refactors the helper into a builder instead, avoiding the need for multiple constructors and empty args