Create execute query callables and wire up api to data client
This uses the ExecuteQueryCallContext to pass the metadata back
to the user using a future. Note that we need to be very careful
here that errors always close this future. That is why watchdog
happens higher in the call chain that for other streaming RPCs

Also note i decided to unwrap executeExceptions so that when
possible the future throws the same exceptions as the row stream

Change-Id: Ie9dd523a980efecb2cff010a2ffd7ac52b5f8ac7